requirements.txt
Flask
pytesseract
Pillow
PyPDF2
python-docx